Convert main/index.js to TypeScript \o/
